pyproject.toml
[tool.poetry]
name = "currency-converter-tgbot"
version = "0.1.0"
description = ""
authors = ["DimaSimakov <dimarodeo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytelegrambotapi = "^4.15.4"
python-dotenv = "^1.0.1"
pytz = "^2024.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
